feat: separation of concers for token fetching and util reorganization
