+++ src/main/java/jumper/model/config/JumperConfig.java
@@ -15,9 +15,9 @@ import io.jsonwebtoken.Jwt;
 import jakarta.validation.constraints.NotNull;
 import java.util.*;
 import jumper.Constants;
-import jumper.service.HeaderUtil;
-import jumper.service.OauthTokenUtil;
-import jumper.service.ObjectMapperUtil;
+import jumper.util.HeaderUtil;
+import jumper.util.OauthTokenUtil;
+import jumper.util.ObjectMapperUtil;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
