+++ src/main/java/jumper/service/TokenGeneratorService.java
@@ -4,17 +4,19 @@
 
 package jumper.service;
 
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.WeakKeyException;
 import java.io.IOException;
 import java.security.GeneralSecurityException;
-import java.util.Date;
-import java.util.HashMap;
+import java.util.*;
+import jumper.Constants;
+import jumper.model.config.JumperConfig;
 import jumper.model.config.KeyInfo;
+import jumper.util.OauthTokenUtil;
 import jumper.util.RsaUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -26,7 +28,7 @@ public class TokenGeneratorService {
 
   private final KeyInfoService keyInfoService;
 
-  public String fromRealm(
+  private String fromRealm(
       HashMap<String, String> claims, String issuer, Date expiration, Date issuedAt) {
     KeyInfo keyInfo;
     try {
@@ -44,7 +46,7 @@ public class TokenGeneratorService {
     return generateToken(claims, issuer, expiration, issuedAt, keyInfo);
   }
 
-  public String fromKey(
+  private String fromKey(
       HashMap<String, String> claims, String issuer, Date expiration, Date issuedAt, String key) {
 
     KeyInfo keyInfo = new KeyInfo();
@@ -56,6 +58,11 @@ public class TokenGeneratorService {
     }
 
     return generateToken(claims, issuer, expiration, issuedAt, keyInfo);
+  }
+
+  public String createJwtTokenFromKey(
+      HashMap<String, String> claims, String issuer, Date expiration, Date issuedAt, String key) {
+    return fromKey(claims, issuer, expiration, issuedAt, key);
   }
 
   private String generateToken(
@@ -75,5 +82,75 @@ public class TokenGeneratorService {
           HttpStatus.UNAUTHORIZED,
           "Key is too weak: The JWT JWA Specification (RFC 7518, Section 3.3) states that keys used with RS256 MUST have a size >= 2048 bits.");
     }
+  }
+
+  public String generateEnhancedLastMileGatewayToken(
+      JumperConfig jc,
+      String operation,
+      String issuer,
+      String publisherId,
+      String subscriberId,
+      boolean legacy) {
+
+    String consumerTokenWithoutSignature =
+        OauthTokenUtil.getTokenWithoutSignature(jc.getConsumerToken());
+
+    Jwt<Header, Claims> consumerTokenClaims =
+        OauthTokenUtil.getAllClaimsFromToken(consumerTokenWithoutSignature);
+
+    Date issuedAt = consumerTokenClaims.getBody().getIssuedAt();
+    Date expiration = consumerTokenClaims.getBody().getExpiration();
+    String sub = consumerTokenClaims.getBody().get(Constants.TOKEN_CLAIM_SUB, String.class);
+    String aud = consumerTokenClaims.getBody().get(Constants.TOKEN_CLAIM_AUD, String.class);
+
+    HashMap<String, String> claims = new HashMap<>();
+    claims.put(Constants.TOKEN_CLAIM_TYP, "Bearer");
+    claims.put(Constants.TOKEN_CLAIM_AZP, "stargate");
+    claims.put(Constants.TOKEN_CLAIM_SUB, sub);
+    claims.put(Constants.TOKEN_CLAIM_REQUEST_PATH, jc.getRequestPath());
+    claims.put(Constants.TOKEN_CLAIM_OPERATION, operation);
+    claims.put(Constants.TOKEN_CLAIM_CLIENT_ID, jc.getConsumer());
+    claims.put(Constants.TOKEN_CLAIM_ORIGIN_ZONE, jc.getConsumerOriginZone());
+    claims.put(Constants.TOKEN_CLAIM_ORIGIN_STARGATE, jc.getConsumerOriginStargate());
+
+    if (legacy) {
+      String consumerTokenSignature = OauthTokenUtil.getSignature(jc.getConsumerToken());
+      claims.put(Constants.TOKEN_CLAIM_ACCESS_TOKEN_SIGNATURE, consumerTokenSignature);
+
+    } else {
+      claims.put(Constants.TOKEN_CLAIM_ACCESS_TOKEN_ENVIRONMENT, jc.getEnvName());
+
+      if (Objects.nonNull(jc.getSecurityScopes())) {
+        claims.put(Constants.TOKEN_CLAIM_SCOPE, jc.getSecurityScopes());
+      }
+
+      if (Objects.nonNull(publisherId)) {
+        claims.put(Constants.TOKEN_CLAIM_ACCESS_TOKEN_PUBLISHER_ID, publisherId);
+      }
+
+      if (Objects.nonNull(subscriberId)) {
+        claims.put(Constants.TOKEN_CLAIM_ACCESS_TOKEN_SUBSCRIBER_ID, subscriberId);
+        claims.put(Constants.TOKEN_CLAIM_AUD, subscriberId);
+      }
+    }
+
+    if (StringUtils.isNotBlank(aud)) {
+      claims.put(Constants.TOKEN_CLAIM_AUD, aud);
+    }
+
+    return fromRealm(claims, issuer, expiration, issuedAt);
+  }
+
+  public String generateGatewayTokenForPublisher(String issuer, String realm) {
+    HashMap<String, String> claims = new HashMap<>();
+    claims.put(Constants.TOKEN_CLAIM_TYP, "Bearer");
+    claims.put(Constants.TOKEN_CLAIM_AZP, "stargate");
+    claims.put(Constants.TOKEN_CLAIM_CLIENT_ID, "gateway");
+
+    return fromRealm(
+        claims,
+        issuer,
+        new Date(System.currentTimeMillis() + 300 * 1000),
+        new Date(System.currentTimeMillis()));
   }
 }
