+++ src/test/java/jumper/mocks/MockIrisServer.java
@@ -17,8 +17,8 @@ import java.util.Base64;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import jumper.model.TokenInfo;
-import jumper.service.ObjectMapperUtil;
 import jumper.util.AccessToken;
+import jumper.util.ObjectMapperUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.mockserver.client.MockServerClient;
 import org.mockserver.integration.ClientAndServer;
